refactor: Restructure internal utils module
