+++ django_slack_tools/slack_messages/admin/message.py
@@ -10,7 +10,7 @@ from django.utils.html import format_html
 from django.utils.translation import gettext_lazy as _
 
 from django_slack_tools.slack_messages.models import SlackMessage
-from django_slack_tools.utils.widgets import JSONWidget
+from django_slack_tools.utils.django.widgets import JSONWidget
 
 if TYPE_CHECKING:
     from django_stubs_ext import StrOrPromise
