+++ django_slack_tools/slack_messages/admin/messaging_policy.py
@@ -12,8 +12,8 @@ from django.utils.html import format_html
 from django.utils.translation import gettext_lazy as _
 
 from django_slack_tools.slack_messages.models import SlackMessagingPolicy
+from django_slack_tools.utils.django.widgets import JSONWidget
 from django_slack_tools.utils.slack import get_block_kit_builder_url
-from django_slack_tools.utils.widgets import JSONWidget
 
 if TYPE_CHECKING:
     from django.http import HttpRequest
